+++ Tests/Carbon.PermissionsTestHelper/Carbon.PermissionsTestHelper.psm1
@@ -1,14 +1,2 @@
 
 
-function Invoke-TestCCryptoKeyAvailable
-{
-    [CmdletBinding()]
-    param(
-    )
-
-    InModuleScope 'Carbon.Permissions' -ScriptBlock {
-        param(
-        )
-        Test-CCryptoKeyAvailable
-    } -Parameters $PSBoundParameters
-}